requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.10.3
python-dotenv==1.1.0
langgraph==1.1.6
faiss-cpu==1.13.2
sentence-transformers==5.3.0
openai==2.30.0
requests==2.33.1
aiohttp==3.11.10
jinja2==3.1.6
python-multipart==0.0.9
